Launcher3: Do not attach cleanupScreenshot() to frame update event
When the animation get canceled, cleanupScreenshot() should
be run instantly because it does something like removing
pending animations. The switchToScreenshot() attaches it to
RtFrameCallback which only be called on frame update. This
is leading to problems as the next frame update may only
happened when the next gesture starts, which means all the
pending animations will be removed on next gesture. And then
the next gesture freezed.

Fixes randomly gesture freeze since 12L.

Change-Id: I10247294a2dcae467706c434685b299f8b525888
Signed-off-by: LibXZR <i@xzr.moe>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
Signed-off-by: Sipun Ku Mahanta <sipunkumar85@gmail.com>
Signed-off-by: Hưng Phan <phandinhhungvp2001@gmail.com>